openmanufacturing/pyproject.toml
[project]
name = "openmanufacturing"
version = "1.0.0"
description = "Advanced optical packaging and assembly platform for photonics manufacturing"
authors = [
    {name = "LlamaFarms Team",email = "team@llamafarms.ai"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "openmanufacturing"
version = "1.0.0"
description = "Advanced optical packaging and assembly platform for photonics manufacturing"
authors = ["OpenManufacturing Team <team@openmanufacturing.org>"]
readme = "README.md"
license = "MIT"
packages = [{include = "openmanufacturing", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
pydantic = "^2.3.0"
opencv-python = "^4.8.0"
numpy = "^1.25.2"
pillow = "^10.0.0"
redis = "^5.0.0"
prometheus-client = "^0.17.1"
pyyaml = "^6.0.1"
typer = "^0.9.0"
rich = "^13.5.2"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"
cryptography = "^41.0.3"
httpx = "^0.24.1"
pandas = "^2.1.0"
matplotlib = "^3.8.0"
scikit-learn = "^1.3.0"
scipy = "^1.11.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
faker = "^19.3.1"
pytest-mock = "^3.11.1"
types-pyyaml = "^6.0.12.11"
types-redis = "^4.6.0.3"
sphinx = "^7.2.5"
sphinx-rtd-theme = "^1.3.0"
pytest-xdist = "^3.3.1"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]
testpaths = ["tests"]
